0-whatsmyname.c

+++ 0x0B-malloc_free/0-whatsmyname.c
@@ -6,7 +6,7 @@
  * @argv: argument vector
  * Return: always 0.
  */
-float main(float argc __attribute__((unused)), char **argv)
+int main(int argc __attribute__((unused)), char **argv)
 {
 	printf("%s\n", *argv);
 	return (0);
